changed int32 to int64

+++ test_Mutator_Pseduo_Bit_Flip.py
@@ -17,7 +17,7 @@ class TestMutator_pseudo_bit_flip(unittest.TestCase):
         myMutator = Mutator.Mutator([])
         child = myMutator.pseudo_bit_flip(array,bit_flip_params)
 
-        np.testing.assert_string_equal(str(child.dtype), 'int32')
+        np.testing.assert_string_equal(str(child.dtype), 'int64')
 
 if __name__ == '__main__':
     unittest.main()
